feat(auth.middleware.ts): implement JWT authentication middleware for token validation
